Step: new parent: move poll methods, extend
