+++ parallel-consumer-core/src/main/java/io/confluent/parallelconsumer/ParentParallelEoSStreamProcessor.java
@@ -18,12 +18,9 @@ import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.apache.kafka.clients.consumer.KafkaConsumer;
 import org.apache.kafka.clients.consumer.MockConsumer;
 import org.apache.kafka.clients.consumer.internals.ConsumerCoordinator;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.clients.producer.RecordMetadata;
 import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.utils.Time;
 import org.slf4j.MDC;
-import pl.tlinkowski.unij.api.UniLists;
 
 import javax.naming.InitialContext;
 import javax.naming.NamingException;
@@ -42,7 +39,6 @@ import static io.confluent.csid.utils.BackportUtils.isEmpty;
 import static io.confluent.csid.utils.BackportUtils.toSeconds;
 import static io.confluent.csid.utils.StringUtils.msg;
 import static io.confluent.parallelconsumer.internal.State.*;
-import static io.confluent.parallelconsumer.internal.UserFunctions.carefullyRun;
 import static java.time.Duration.ofMillis;
 import static java.util.concurrent.TimeUnit.MILLISECONDS;
 import static java.util.concurrent.TimeUnit.SECONDS;
@@ -52,7 +48,7 @@ import static lombok.AccessLevel.PROTECTED;
  * @see ParallelConsumer
  */
 @Slf4j
-public class ParentParallelEoSStreamProcessor<K, V> implements ParallelStreamProcessor<K, V>, ConsumerRebalanceListener, Closeable {
+public abstract class ParentParallelEoSStreamProcessor<K, V> implements ParallelStreamProcessor<K, V>, ConsumerRebalanceListener, Closeable {
 
     public static final String MDC_INSTANCE_ID = "pcId";
 
@@ -358,78 +354,6 @@ public class ParentParallelEoSStreamProcessor<K, V> implements ParallelStreamPro
         } else {
             // noop - probably MockConsumer being used in testing - which doesn't do auto commits
         }
-    }
-
-    @Override
-    public void poll(Consumer<ConsumerRecord<K, V>> usersVoidConsumptionFunction) {
-        Function<ConsumerRecord<K, V>, List<Object>> wrappedUserFunc = (record) -> {
-            log.trace("asyncPoll - Consumed a record ({}), executing void function...", record.offset());
-
-            carefullyRun(usersVoidConsumptionFunction, record);
-
-            log.trace("asyncPoll - user function finished ok.");
-            return UniLists.of(); // user function returns no produce records, so we satisfy our api
-        };
-        Consumer<Object> voidCallBack = (ignore) -> log.trace("Void callback applied.");
-        supervisorLoop(wrappedUserFunc, voidCallBack);
-    }
-
-    @Override
-    @SneakyThrows
-    public void pollAndProduceMany(Function<ConsumerRecord<K, V>, List<ProducerRecord<K, V>>> userFunction,
-                                   Consumer<ConsumeProduceResult<K, V, K, V>> callback) {
-        // todo refactor out the producer system to a sub class
-        if (!options.isProducerSupplied()) {
-            throw new IllegalArgumentException("To use the produce flows you must supply a Producer in the options");
-        }
-
-        // wrap user func to add produce function
-        Function<ConsumerRecord<K, V>, List<ConsumeProduceResult<K, V, K, V>>> wrappedUserFunc = (consumedRecord) -> {
-
-            List<ProducerRecord<K, V>> recordListToProduce = carefullyRun(userFunction, consumedRecord);
-
-            if (recordListToProduce.isEmpty()) {
-                log.debug("No result returned from function to send.");
-            }
-            log.trace("asyncPoll and Stream - Consumed a record ({}), and returning a derivative result record to be produced: {}", consumedRecord, recordListToProduce);
-
-            List<ConsumeProduceResult<K, V, K, V>> results = new ArrayList<>();
-            log.trace("Producing {} messages in result...", recordListToProduce.size());
-            for (ProducerRecord<K, V> toProduce : recordListToProduce) {
-                log.trace("Producing {}", toProduce);
-                RecordMetadata produceResultMeta = producerManager.get().produceMessage(toProduce);
-                var result = new ConsumeProduceResult<>(consumedRecord, toProduce, produceResultMeta);
-                results.add(result);
-            }
-            return results;
-        };
-
-        supervisorLoop(wrappedUserFunc, callback);
-    }
-
-    @Override
-    @SneakyThrows
-    public void pollAndProduceMany(Function<ConsumerRecord<K, V>, List<ProducerRecord<K, V>>> userFunction) {
-        pollAndProduceMany(userFunction, (record) -> {
-            // no op call back
-            log.trace("No-op user callback");
-        });
-    }
-
-    @Override
-    @SneakyThrows
-    public void pollAndProduce(Function<ConsumerRecord<K, V>, ProducerRecord<K, V>> userFunction) {
-        pollAndProduce(userFunction, (record) -> {
-            // no op call back
-            log.trace("No-op user callback");
-        });
-    }
-
-    @Override
-    @SneakyThrows
-    public void pollAndProduce(Function<ConsumerRecord<K, V>, ProducerRecord<K, V>> userFunction,
-                               Consumer<ConsumeProduceResult<K, V, K, V>> callback) {
-        pollAndProduceMany((record) -> UniLists.of(userFunction.apply(record)), callback);
     }
 
     /**
